+++ client/src/tests/fresh.rs
@@ -9,7 +9,7 @@ use crate::Location;
 
 /// Generates a random [`Location`].
 pub fn location() -> Location {
-    Location::generic(bytestring(4096), bytestring(4096))
+    Location::generic(variable_bytestring(4096), variable_bytestring(4096))
 }
 
 /// Creates a random hd_path.
@@ -18,7 +18,7 @@ pub fn hd_path() -> (String, Chain) {
     let mut is = vec![];
     while coinflip() {
         let i = random::<u32>() & 0x7fffff;
-        write!(&mut s, "/{}", i).expect("");
+        write!(&mut s, "/{}'", i).expect("Failed appending path segment");
         is.push(i);
     }
     (s, Chain::from_u32_hardened(is))