+++ engine/fuzz/src/vault_fuzz.rs
@@ -1,5 +1,6 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
+
 #![no_main]
 
 use engine::vault::{DbView, Key, RecordHint, RecordId, VaultId};